registration request classes moved to dto package
