+++ src/main/java/com/msn/msncars/auth/AuthService.java
@@ -1,5 +1,9 @@
 package com.msn.msncars.auth;
 
+import com.msn.msncars.auth.dto.CompanyRegistrationRequest;
+import com.msn.msncars.auth.dto.CompanyRegistrationResponse;
+import com.msn.msncars.auth.dto.UserRegistrationRequest;
+
 public interface AuthService {
     String registerUserAndAssignRole(UserRegistrationRequest userRegistrationRequest, AccountRole accountRole);
     CompanyRegistrationResponse registerCompany(CompanyRegistrationRequest companyRegistrationRequest);
