+++ src/main/java/com/msn/msncars/auth/dto/CompanyRegistrationResponse.java
@@ -1,4 +1,4 @@
-package com.msn.msncars.auth;
+package com.msn.msncars.auth.dto;
 
 public record CompanyRegistrationResponse(String userId, Long companyId) {
 }
