+++ src/main/java/com/msn/msncars/auth/dto/UserRegistrationRequest.java
@@ -1,3 +1,3 @@
-package com.msn.msncars.auth;
+package com.msn.msncars.auth.dto;
 
 public record UserRegistrationRequest(String username, String password, String email, String firstName, String lastName) { }
